+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Fixed
+
+* Correction des paramètres d'obtention de la liste des dossiers
+* Utilisation de \Lizmap\Request\Proxy::getRemoteData à la place de curl
+
 ## 1.0.1 - 2023-10-26
 
 ### Added